requirements.txt
smolagents
streamlit
sentence-transformers
qdrant-client
torch
mem0ai
dotenv